feature: created controllers for the game
Made controllers for each card component and game component. This is
needed for the views to interact with the models. After weighing some
design decisions, the class State is shared between game and view. This
was to avoid duplication and for the scope of this project, I do not
expect their usages in the view and model components to differ.