refactor: Completely removed GameLogic in Game
Added GameEngine to Game to remove the logic responsibility in Game.
This led to several refactorings as a result. Now, GameEngineImpl holds
the state of gameContext and mutates it. The reason for this is because
I did not want to expose unnecessary components of that GameEngine would
use. Additionally, from the refactoring, it seems reasonable for
GameEngine to hold the responsibility of determining if a game is
completed. As a result, GameEngineImpl now determines how a game is
completed.

Also, I created factory methods for commonly used tasks.