chore: changed build work flow to webpack
The motivation for this was to create a simple way to create a web app
for both production and development. After some research, I though the
best way to configure the build process for my projects is using
webpack. The main objective is to create a clean script for both dev and
prod. As a result, I tried to make the process as modular as possible by
breaking the components into separate files. I first created a baseline
for the dev.